Moved all `rustc` internal <=> `c2rust_analysis_rt` type conversions to `crate::runtime_conversions`.
